urun_iade_risk_skoru/requirements.txt
tensorflow>=2.15.0
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.24.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0 